Changed the way pieces are tracked
Instances of men are now stored in an array in the Board node and are no longer tracked seperately.

Checking if you are trying to move to an empty space is now done on input instead of on move validation.

Selecting a piece no longer uses the piece's CollisionShape2D. Instead you select a piece by clicking on it's tile on the Board.